[pkg/otlp/model] Move NaN metric test to its own file
