+++ pkg/otlp/model/translator/metrics_translator_test.go
@@ -25,10 +25,7 @@ import (
 	"go.opentelemetry.io/collector/pdata/pcommon"
 	"go.opentelemetry.io/collector/pdata/pmetric"
 	"go.uber.org/zap"
-	"go.uber.org/zap/zapcore"
-	"go.uber.org/zap/zaptest/observer"
 
-	"github.com/DataDog/datadog-agent/pkg/otlp/model/attributes"
 	"github.com/DataDog/datadog-agent/pkg/otlp/model/source"
 	"github.com/DataDog/datadog-agent/pkg/quantile"
 	"github.com/DataDog/datadog-agent/pkg/quantile/summary"
@@ -650,32 +647,6 @@ func TestFormatFloat(t *testing.T) {
 	}
 }
 
-func exampleSummaryDataPointSlice(ts pcommon.Timestamp, sum float64, count uint64) pmetric.SummaryDataPointSlice {
-	slice := pmetric.NewSummaryDataPointSlice()
-	point := slice.AppendEmpty()
-	point.SetCount(count)
-	point.SetSum(sum)
-	qSlice := point.QuantileValues()
-
-	qMin := qSlice.AppendEmpty()
-	qMin.SetQuantile(0.0)
-	qMin.SetValue(0)
-
-	qMedian := qSlice.AppendEmpty()
-	qMedian.SetQuantile(0.5)
-	qMedian.SetValue(100)
-
-	q999 := qSlice.AppendEmpty()
-	q999.SetQuantile(0.999)
-	q999.SetValue(500)
-
-	qMax := qSlice.AppendEmpty()
-	qMax.SetQuantile(1)
-	qMax.SetValue(600)
-	point.SetTimestamp(ts)
-	return slice
-}
-
 const (
 	testHostname     = "res-hostname"
 	fallbackHostname = "fallbackHostname"
@@ -737,124 +708,4 @@ func newSketchWithHostname(name string, summary summary.Summary, tags []string) 
 	s := newSketch(dims.AddTags(tags...), 0, summary)
 	s.host = testHostname
 	return s
-}
-
-func createNaNMetrics() pmetric.Metrics {
-	md := pmetric.NewMetrics()
-	rms := md.ResourceMetrics()
-	rm := rms.AppendEmpty()
-
-	attrs := rm.Resource().Attributes()
-	attrs.PutStr(attributes.AttributeDatadogHostname, testHostname)
-	ilms := rm.ScopeMetrics()
-
-	metricsArray := ilms.AppendEmpty().Metrics()
-
-	// DoubleGauge
-	met := metricsArray.AppendEmpty()
-	met.SetName("nan.gauge")
-	met.SetEmptyGauge()
-	dpsDouble := met.Gauge().DataPoints()
-	dpDouble := dpsDouble.AppendEmpty()
-	dpDouble.SetTimestamp(seconds(0))
-	dpDouble.SetDoubleValue(math.NaN())
-
-	// Double Sum (delta)
-	met = metricsArray.AppendEmpty()
-	met.SetName("nan.delta.sum")
-	met.SetEmptySum()
-	met.Sum().SetAggregationTemporality(pmetric.AggregationTemporalityDelta)
-	dpsDouble = met.Sum().DataPoints()
-	dpDouble = dpsDouble.AppendEmpty()
-	dpDouble.SetTimestamp(seconds(0))
-	dpDouble.SetDoubleValue(math.NaN())
-
-	// Double Sum (delta monotonic)
-	met = metricsArray.AppendEmpty()
-	met.SetName("nan.delta.monotonic.sum")
-	met.SetEmptySum()
-	met.Sum().SetAggregationTemporality(pmetric.AggregationTemporalityDelta)
-	dpsDouble = met.Sum().DataPoints()
-	dpDouble = dpsDouble.AppendEmpty()
-	dpDouble.SetTimestamp(seconds(0))
-	dpDouble.SetDoubleValue(math.NaN())
-
-	// Histogram
-	met = metricsArray.AppendEmpty()
-	met.SetName("nan.histogram")
-	met.SetEmptyHistogram()
-	met.Histogram().SetAggregationTemporality(pmetric.AggregationTemporalityDelta)
-	dpsDoubleHist := met.Histogram().DataPoints()
-	dpDoubleHist := dpsDoubleHist.AppendEmpty()
-	dpDoubleHist.SetCount(20)
-	dpDoubleHist.SetSum(math.NaN())
-	dpDoubleHist.BucketCounts().FromRaw([]uint64{2, 18})
-	dpDoubleHist.ExplicitBounds().FromRaw([]float64{0})
-	dpDoubleHist.SetTimestamp(seconds(0))
-
-	// Double Sum (cumulative)
-	met = metricsArray.AppendEmpty()
-	met.SetName("nan.cumulative.sum")
-	met.SetEmptySum()
-	met.Sum().SetAggregationTemporality(pmetric.AggregationTemporalityCumulative)
-	dpsDouble = met.Sum().DataPoints()
-	dpsDouble.EnsureCapacity(2)
-	dpDouble = dpsDouble.AppendEmpty()
-	dpDouble.SetTimestamp(seconds(0))
-	dpDouble.SetDoubleValue(math.NaN())
-
-	// Double Sum (cumulative monotonic)
-	met = metricsArray.AppendEmpty()
-	met.SetName("nan.cumulative.monotonic.sum")
-	met.SetEmptySum()
-	met.Sum().SetAggregationTemporality(pmetric.AggregationTemporalityCumulative)
-	met.Sum().SetIsMonotonic(true)
-	dpsDouble = met.Sum().DataPoints()
-	dpsDouble.EnsureCapacity(2)
-	dpDouble = dpsDouble.AppendEmpty()
-	dpDouble.SetTimestamp(seconds(0))
-	dpDouble.SetDoubleValue(math.NaN())
-
-	// Summary
-	met = metricsArray.AppendEmpty()
-	met.SetName("nan.summary")
-	met.SetEmptySummary()
-	slice := exampleSummaryDataPointSlice(seconds(0), math.NaN(), 1)
-	slice.CopyTo(met.Summary().DataPoints())
-
-	met = metricsArray.AppendEmpty()
-	met.SetName("nan.summary")
-	met.SetEmptySummary()
-	slice = exampleSummaryDataPointSlice(seconds(2), 10_001, 101)
-	slice.CopyTo(met.Summary().DataPoints())
-	return md
-}
-
-func TestNaNMetrics(t *testing.T) {
-	md := createNaNMetrics()
-
-	core, observed := observer.New(zapcore.DebugLevel)
-	testLogger := zap.New(core)
-	ctx := context.Background()
-	tr := newTranslator(t, testLogger)
-	consumer := &mockFullConsumer{}
-	err := tr.MapMetrics(ctx, md, consumer)
-	require.NoError(t, err)
-
-	assert.ElementsMatch(t, consumer.metrics, []metric{
-		newCountWithHostname("nan.summary.count", 100, 2, []string{}),
-	})
-
-	assert.ElementsMatch(t, consumer.sketches, []sketch{
-		newSketchWithHostname("nan.histogram", summary.Summary{
-			Min: 0,
-			Max: 0,
-			Sum: 0,
-			Avg: 0,
-			Cnt: 20,
-		}, []string{}),
-	})
-
-	// One metric type was unknown or unsupported
-	assert.Equal(t, observed.FilterMessage("Unsupported metric value").Len(), 7)
 }
